+++ buffer.h
@@ -1,11 +1,11 @@
 typedef struct {
-    int8_t         size;   /* maximum number of elements           */
-    int8_t         start;  /* index of oldest element              */
-    int8_t         end;    /* index at which to write new element  */
+    int16_t         size;   /* maximum number of elements           */
+    int16_t         start;  /* index of oldest element              */
+    int16_t         end;    /* index at which to write new element  */
     char        *elems;  /* vector of elements                   */
 } CircularBuffer;
 
-void cbInit(CircularBuffer *cb, int8_t size);
+void cbInit(CircularBuffer *cb, int16_t size);
 
 void cbFree(CircularBuffer *cb);
